Wake up babe time for cleaning
less code and better readability (i guess) also extracted the Log method now it's in its own module this will make it easier when implemanting the GUI later
